feat(*): work as package
