--- bundles/index/index.bemdecl.js
@@ -1,5 +1,0 @@
-exports.blocks = [
-  {
-    "name": "root"
-  }
-];
