+++ public/config.json
@@ -1,0 +1,26 @@
+{
+  "bb": true,
+  "title": {
+    "ru": "Блог",
+    "en": "Blog"
+  },
+  "subtitle": {
+    "ru": "на стеке БЭМ-технологий",
+    "en": "on BEM stack"
+	},
+  "langs": ["ru", "en"],
+  "defaultLang": "en",
+  "layouts": {
+    "blog": {
+      "paginate": 10
+    }
+  },
+  "theme": "simple",
+  "output": "dist",
+  "users": [
+    {
+      "login": "tadatuta",
+      "email": "i@tadatuta.com"
+    }
+  ]
+}
