+++ public/themes/simple/footer/footer.css
@@ -1,0 +1,3 @@
+.footer {
+  height: 40px;
+}
