+++ public/themes/simple/header/header.deps.js
@@ -1,0 +1,5 @@
+({
+	shouldDeps: [
+		{ block: 'language-switcher' }
+	]
+});
