+++ public/themes/simple/language-switcher/language-switcher.css
@@ -1,0 +1,4 @@
+.language-switcher {
+	font-size: 20px;
+	float: left;
+}
