--- scripts/build
@@ -1,14 +1,0 @@
-OUTPUT_DIR=./dist
-CSS_DIR=$OUTPUT_DIR/css
-JS_DIR=$OUTPUT_DIR/js
-
-rm -rf $OUTPUT_DIR
-YENV=production enb make
-node scripts/grab.js
-node scripts/generate.js
-mkdir $CSS_DIR
-mkdir $JS_DIR
-cp bundles/index/index.min.css $CSS_DIR/
-cp bundles/index/index.min.js $JS_DIR/
-cp favicon.ico $OUTPUT_DIR/
-touch $OUTPUT_DIR/.nojekyll
