--- scripts/grab.js
@@ -1,25 +1,0 @@
-'use strict';
-
-const fs = require('bluebird').promisifyAll(require('fs-extra'));
-const {join} = require('path');
-const marked = require('meta-marked');
-const config = require('../src/content/config.json');
-const mdFiles = require('glob').sync('src/content/*/*.md'); // TODO: get from config
-
-const results = mdFiles.map(file => {
-  const md = fs.readFileSync(file, 'utf-8');
-  const parsed = marked(md);
-  const fileName = file.split('/').reverse()[0];
-
-  return {
-    fileName: fileName,
-    name: fileName.split('.')[0],
-    layout: file.split('/').reverse()[1],
-    path: file,
-    lang: file.split('.').reverse()[1],
-    meta: parsed.meta,
-    content: parsed.html // TODO: posthtml?
-  };
-});
-
-fs.outputFileSync(join('dist', 'data.json'), JSON.stringify(results));
