--- src/components/content/_type/content_type_list.bemtree.js
@@ -1,24 +1,0 @@
-block('content').mod('type', 'list').content()(function() {
-  return this._data.map(function(post) {
-    return [
-      {
-        block: 'post',
-        content: [
-          {
-            tag: 'h2',
-            content: {
-              block: 'link',
-              url: post.fileName.replace('md', 'html'),
-              content: post.meta.title
-            }
-          },
-          post.content,
-          {
-            block: 'tags',
-            content: post.meta && post.meta.tags
-          }
-        ]
-      }
-    ];
-  });
-});
