--- src/components/content/content.deps.js
@@ -1,8 +1,0 @@
-({
-  shouldDeps: [
-    { mods: { type: ['single', 'list'] } },
-    'post',
-    'tags',
-    'link'
-  ]
-})
