+++ src/components/page/page.bemtree.js
@@ -1,0 +1,8 @@
+block('page')(
+  content()(function() {
+    return [{
+      block: this.mods.layout,
+      pagination: this._pagination
+    }];
+  })
+);
