+++ src/components/page/page.deps.js
@@ -1,0 +1,9 @@
+({
+  shouldDeps: [
+    { elems: ['meta', 'css', 'js'] },
+    { block: 'blog' },
+    { block: 'simple' },
+    { block: 'post' },
+    { block: 'tags' }
+  ]
+});
