--- src/components/tags/tags.bemtree.js
@@ -1,13 +1,0 @@
-block('tags').content()(function() {
-  var tags = applyNext();
-  return tags.map(function(tag, idx) {
-    return [
-      {
-        block: 'link',
-        url: 'tag-' + tag + '.' + this._lang + '.html',
-        content: tag
-      },
-      tags.length > idx + 1 ? ', ' : ''
-    ];
-  }, this);
-});
