--- src/content/config.json
@@ -1,24 +1,0 @@
-{
-  "title": {
-    "ru": "Блог",
-    "en": "Blog"
-  },
-  "subtitle": {
-    "ru": "на стеке БЭМ-технологий",
-    "en": "on BEM stack"
-	},
-  "langs": ["ru", "en"],
-  "defaultLang": "en",
-  "layouts": {
-    "blog": {
-      "paginate": 10
-    }
-  },
-  "outputFolder": "dist",
-  "users": [
-    {
-      "login": "tadatuta",
-      "email": "i@tadatuta.com"
-    }
-  ]
-}
