--- src/layouts/page/_type/page_type_tags.bemtree.js
@@ -1,9 +1,0 @@
-block('page').mod('type', 'tags').content()(function() {
-    return this.ctx.content.map(function(tag) {
-        return {
-            block: 'link',
-            url: 'tag-' + tag + this._lang + '.html',
-            content: tag
-        };
-    });
-});
