--- src/layouts/page/page.deps.js
@@ -1,10 +1,0 @@
-({
-  shouldDeps: [
-    { elems: ['meta', 'css', 'js'] },
-    { mods: { type: 'tags' } },
-    'link',
-    'header',
-    'content',
-    'footer'
-  ]
-})
