+++ src/layouts/simple/simple.deps.js
@@ -1,0 +1,7 @@
+({
+  shouldDeps: [
+    { block: 'link' },
+    { block: 'header' },
+    { block: 'footer' }
+  ]
+});
