--- src/themes/simple/avatar/avatar.bemhtml.js
@@ -1,6 +1,0 @@
-block('avatar')(
-  tag()('img'),
-  attrs()(function() {
-    return { src: this.ctx.url };
-  })
-)
