--- src/themes/simple/avatar/avatar.css
@@ -1,15 +1,0 @@
-.avatar
-{
-    position: absolute;
-    right: -110px;
-    top: -45px;
-
-    overflow: hidden;
-    border-radius: 50%;
-    border: 3px solid #fff;
-
-    width: 150px;
-    height: 150px;
-    /*background: url(https://d262ilb51hltx0.cloudfront.net/fit/c/200/200/0*-N6wxNecet4yOR92.jpg) no-repeat;
-    background-size: 100%;*/
-}
