--- src/themes/simple/content/content.css
@@ -1,3 +1,0 @@
-.content
-{
-}
