--- src/themes/simple/footer/footer.css
@@ -1,5 +1,0 @@
-/*  TODO: remove me */
-.footer
-{
-    height: 40px;
-}
