--- src/themes/simple/header/header.bemtree.js
@@ -1,9 +1,0 @@
-block('header').content()(function() {
-  return [
-    {
-      block: 'avatar',
-      url: this.bb.avatar
-    },
-    applyNext()
-  ];
-});
