--- src/themes/simple/header/header.deps.js
@@ -1,3 +1,0 @@
-({
-	shouldDeps: ['language-switcher', 'avatar']
-})
