--- src/themes/simple/language-switcher/language-switcher.css
@@ -1,5 +1,0 @@
-.language-switcher
-{
-	font-size: 20px;
-	float: left;
-}
