--- src/themes/simple/link/link.css
@@ -1,12 +1,0 @@
-a,
-.link
-{
-    color: #444;
-    text-decoration: none;
-}
-
-a:hover,
-.link:hover
-{
-    color: #09b;
-}
