--- src/themes/simple/page/page.bemhtml.js
@@ -1,6 +1,0 @@
-block('page').content()(function() {
-  return {
-    elem: 'inner',
-    content: applyNext()
-  };
-});
