--- src/themes/simple/page/page.css
@@ -1,14 +1,0 @@
-.page
-{
-    font-family: 'Gentium Book Basic', Georgia, serif;
-    color: #444;
-    background: #ffba00;
-    margin: 0;
-    padding-top: 150px;
-}
-
-.page__inner
-{
-    width: 750px;
-    margin: auto;
-}
