--- src/themes/simple/post/post.css
@@ -1,12 +1,0 @@
-.post h2
-{
-    font-weight: normal;
-    font-size: 50px;
-}
-
-.post p
-{
-    font-size: 22px;
-    line-height: 1.5;
-    margin-bottom: 30px;
-}
